+++ index.ts
@@ -1,15 +1,18 @@
 import { relative } from "path";
 import {
 	Application,
+	CommentDisplayPart,
 	ContainerReflection,
 	Context,
 	Converter,
 	DeclarationReflection,
+	InlineTagDisplayPart,
 	JSX,
 	ParameterType,
 	ProjectReflection,
 	Reflection,
 	ReflectionKind,
+	ReflectionSymbolId,
 	Renderer,
 	TypeScript as ts,
 } from "typedoc";
@@ -46,6 +49,7 @@ export function load(app: Application) {
 	const referencedSymbols = new Map<ts.Program, Set<ts.Symbol>>();
 	const symbolToOwningModule = new Map<ts.Symbol, Reflection>();
 	const knownPrograms = new Map<Reflection, ts.Program>();
+	const symbolIdKeyToSymbol = new Map<string, ts.Symbol>();
 
 	/**
 	 * Extracts the set of referenced but export-missing symbols owned by a particular module.
@@ -103,11 +107,12 @@ export function load(app: Application) {
 
 	const origCreateSymbolId = Context.prototype.createSymbolId;
 	Context.prototype.createSymbolId = function(symbol, declaration) {
-		if (app.options.getValue("includeDocCommentReferences")) {
-			const owningModule = getOwningModule(this);
-			markSymbolReferenced(symbol, owningModule, this.program);
-		}
-		return origCreateSymbolId.call(this, symbol, declaration);
+		// We don't immediately mark the symbol as referenced here because we
+		// don't want to include referenced symbols from comments which will
+		// be excluded by the excludeTags option.
+		const id = origCreateSymbolId.call(this, symbol, declaration);
+		symbolIdKeyToSymbol.set(id.getStableKey(), symbol);
+		return id;
 	};
 
 	app.options.addDeclaration({
@@ -158,6 +163,17 @@ export function load(app: Application) {
 		}
 	});
 
+	function addCommentReferencedSymbols(context: Context, refl: Reflection) {
+		if (app.options.getValue("includeDocCommentReferences")) {
+			visitReflectionLinkTags(refl, app.options.getValue("excludeTags"), part => {
+				const symbol = part.target instanceof ReflectionSymbolId && symbolIdKeyToSymbol.get(part.target.getStableKey());
+				if (symbol) {
+					markSymbolReferenced(symbol, getOwningModule(context), context.program);
+				}
+			});
+		}
+	}
+
 	app.converter.on(
 		Converter.EVENT_CREATE_DECLARATION,
 		(context: Context, refl: Reflection) => {
@@ -183,8 +199,14 @@ export function load(app: Application) {
 					app.options.getValue("basePath") || process.cwd(),
 				);
 			}
+
+			addCommentReferencedSymbols(context, refl);
 		},
 	);
+
+	app.converter.on(Converter.EVENT_CREATE_SIGNATURE, (context, sig) => {
+		addCommentReferencedSymbols(context, sig);
+	});
 
 	app.converter.on(
 		Converter.EVENT_RESOLVE_BEGIN,
@@ -197,6 +219,8 @@ export function load(app: Application) {
 				modules.push(context.project);
 			}
 
+			let first = true;
+
 			for (const mod of modules) {
 				const program = knownPrograms.get(mod);
 				if (!program) continue;
@@ -206,6 +230,13 @@ export function load(app: Application) {
 				// create symbol IDs, which requires a program due to our override of the method
 				// on TypeDoc's Context.
 				context.setActiveProgram(program);
+
+				if (first) {
+					// We have to do this here because we rely on there being an active program,
+					// but there isn't any active program during the CREATE_PROJECT event.
+					addCommentReferencedSymbols(context, context.project);
+					first = false;
+				}
 
 				let missing = discoverMissingExports(mod, context, program);
 				if (!missing.size) continue;
@@ -264,6 +295,7 @@ export function load(app: Application) {
 			knownPrograms.clear();
 			referencedSymbols.clear();
 			symbolToOwningModule.clear();
+			symbolIdKeyToSymbol.clear();
 		},
 		1e9,
 	);
@@ -279,6 +311,49 @@ export function load(app: Application) {
 				}));
 		}
 	});
+}
+
+function visitReflectionLinkTags(
+	reflection: Reflection,
+	excludeTags: readonly `@${string}`[],
+	visitor: (part: InlineTagDisplayPart) => void,
+) {
+	if (reflection.isProject() || reflection.isDeclaration()) {
+		visitLinkTags(reflection.readme || [], visitor);
+	}
+
+	if (reflection.isDocument()) {
+		visitLinkTags(reflection.content || [], visitor);
+	}
+
+	if (reflection.comment) {
+		visitLinkTags(reflection.comment.summary, visitor);
+		for (const tag of reflection.comment.blockTags) {
+			if (!excludeTags.includes(tag.tag)) {
+				visitLinkTags(tag.content, visitor);
+			}
+		}
+	}
+
+	if (
+		reflection.isDeclaration()
+		&& reflection.kindOf(ReflectionKind.TypeAlias)
+		&& reflection.type?.type === "union"
+		&& reflection.type.elementSummaries
+	) {
+		for (const parts of reflection.type.elementSummaries) {
+			visitLinkTags(parts, visitor);
+		}
+	}
+}
+
+function visitLinkTags(parts: readonly CommentDisplayPart[], visitor: (part: InlineTagDisplayPart) => void) {
+	const linkTags = ["@link", "@linkcode", "@linkplain"];
+	for (const part of parts) {
+		if (part.kind === "inline-tag" && linkTags.includes(part.tag)) {
+			visitor(part);
+		}
+	}
 }
 
 function getOwningModule(context: Context): Reflection {